Fase2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
arrow==1.4.0
asttokens==3.0.0
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
billiard==4.2.2
binaryornot==0.4.4
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cookiecutter==2.6.0
cookiecutter-data-science==2.3.0
cryptography==46.0.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.2
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.2.1
-e git+https://github.com/AlejandroDiaz10/MLOps.git@3a113d66c97d85167755ab49be2ff436e1bb8c79#egg=fase2&subdirectory=Fase2
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.8.0
hydra-core==1.3.2
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kombu==5.5.4
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
omegaconf==2.3.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
pyzmq==27.1.0
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.2
scikit-learn==1.7.2
scipy==1.16.2
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
stack-data==0.6.3
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
wcwidth==0.2.14
xgboost==3.0.5
yarl==1.22.0
zc.lockfile==4.0
